+++ CHANGELOG.md
@@ -5,3 +5,32 @@
 <p align="center"><img src="https://assets-cdn.github.com/favicon.ico" width=24 height=24/> <a href="https://github.com/arcticicestudio/styled-modern-normalize/releases/latest"><img src="https://img.shields.io/github/release/arcticicestudio/styled-modern-normalize.svg?style=flat-square"/></a> <img src="https://www.npmjs.com/static/images/touch-icons/favicon-32x32.png" width=24 height=24/> <a href="https://www.npmjs.com/package/styled-modern-normalize"><img src="https://img.shields.io/npm/v/styled-modern-normalize.svg?style=flat-square"/></a> <a href="https://www.npmjs.com/package/styled-modern-normalize"><img src="https://img.shields.io/npm/dt/styled-modern-normalize.svg?style=flat-square"/></a> <a href="https://www.npmjs.com/package/styled-modern-normalize"><img src="https://img.shields.io/npm/dm/styled-modern-normalize.svg?style=flat-square"/></a></p>
 
 <p align="center"><a href="https://github.com/arcticicestudio/styleguide-javascript"><img src="https://img.shields.io/badge/modern--normalize.css-0.4.0-5E81AC.svg?style=flat-square"/></a></p>
+
+# 0.1.0
+
+![Release Date: 2018-05-19](https://img.shields.io/badge/Release_Date-2018--05--19-88C0D0.svg?style=flat-square) [![Project Board](https://img.shields.io/badge/Project_Board-0.1.0-88C0D0.svg?style=flat-square)](https://github.com/arcticicestudio/styled-modern-normalize/projects/3) [![Milestone](https://img.shields.io/badge/Milestone-0.1.0-88C0D0.svg?style=flat-square)](https://github.com/arcticicestudio/styled-modern-normalize/milestone/1)
+
+> Detailed information can be found in the [project documentation][docs].
+
+### Features
+
+❯ Implemented the initial API for the CSS definitions based on [modern-normalize][gh-modern-normalize] version [0.4.0][npm-modern-normalize-v0.4.0].
+
+The package exposes the following API exports. (#1 in PR #2, 8768a8f7)
+
+* default export `modernNormalize`
+* `MODERN_NORMALIZE_VERSION` to provide the currently supported _modern-normalize.css_ version
+* both are also available as named exports
+
+Detailed usage instructions can be found in the [usage documentation][docs-usage].
+
+<p align="center"><img src="https://raw.githubusercontent.com/arcticicestudio/nord/develop/assets/banner-footer-mountains.svg?sanitize=true" /></p>
+
+<p align="center">Copyright &copy; 2018-present Arctic Ice Studio & Sven Greb</p>
+
+<p align="center"><a href="https://github.com/arcticicestudio/styled-modern-normalize/blob/develop/LICENSE.md"><img src="https://img.shields.io/badge/License-MIT-5E81AC.svg?style=flat-square"/></a></p>
+
+[docs]: https://github.com/arcticicestudio/styled-modern-normalize#readme
+[docs-usage]: https://github.com/arcticicestudio/styled-modern-normalize#usage
+[gh-modern-normalize]: https://github.com/sindresorhus/modern-normalize
+[npm-modern-normalize-v0.4.0]: https://www.npmjs.com/package/modern-normalize/v/0.4.0